requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
requests
ultralytics